+++ modules/Eligibility/TeacherCompletion.php
@@ -47,7 +47,8 @@ $periods_RET = DBGet( "SELECT sp.PERIOD_ID,sp.TITLE,COALESCE(sp.SHORT_NAME,sp.TI
 		AND cp.SYEAR='" . UserSyear() . "')
 	ORDER BY sp.SORT_ORDER IS NULL,sp.SORT_ORDER,sp.TITLE" );
 
-$period_select = '<select name="period" id="period"><option value="">' . _( 'All' ) . '</option>';
+$period_select = '<select name="period" id="period" onChange="ajaxPostForm(this.form,true);">
+	<option value="">' . _( 'All' ) . '</option>';
 
 foreach ( (array) $periods_RET as $period )
 {
@@ -80,11 +81,10 @@ if ( $start
 	}
 }
 
-$date_select = '<select name="start_date" id="start_date">' . $date_select . '</select>';
+$date_select = '<select name="start_date" id="start_date" onChange="ajaxPostForm(this.form,true);">' . $date_select . '</select>';
 
 DrawHeader( '<label for="start_date">' . _( 'Timeframe' ) . ':</label> ' . $date_select . ' &mdash; ' .
-	'<label for="period">' . _( 'Period' ) . ':</label> ' . $period_select . ' ' .
-	SubmitButton( _( 'Go' ) ) );
+	'<label for="period">' . _( 'Period' ) . ':</label> ' . $period_select );
 
 echo '</form>';