+++ CHANGES.md
@@ -10,6 +10,7 @@ Changes in 10.9
 - Exclude already associated parents/students from Search() in AddUsers.php & AddStudents.php
 - Replace "Course Title" in TipMessage with actual Course title in Attendance/, Eligibility/ & Grades/TeacherCompletion.php
 - Handle multiple Course Periods on the same Period for same Teacher in Attendance/ & Eligibility/TeacherCompletion.php
+- JS Remove "Go" button & submit form on select change in Eligibility/TeacherCompletion.php
 
 Changes in 10.8.4
 -----------------